+++ packages/modules/typsert/test/TEST-UTILS.ts
@@ -1,0 +1,9 @@
+/**
+ * Convenience strict checks
+ */
+export declare function IsTrue<T extends true>(params?: T): T;
+
+/**
+ * Convenience strict checks
+ */
+export declare function IsFalse<T extends false>(params?: T): T;